+++ api/prisma/migrations/20260212143551_add_endpoint_log_config/migration.sql
@@ -1,2 +1,0 @@
--- AlterTable
-ALTER TABLE "Endpoint" ADD COLUMN "logConfig" TEXT;